% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_questions.R
\docType{data}
\name{all_questions}
\alias{all_questions}
\title{All trivia questions that are used in the app}
\format{
A data frame with 49717 rows and 4 variables:
\describe{
  \item{question}{a trivia question}
  \item{category}{the trivia category the question belongs to}
  \item{answer}{the correct answer to the question}
  \item{choices}{allowed answer choices for the question}
}
}
\source{
\url{https://github.com/uberspot/OpenTriviaQA}
}
\usage{
all_questions
}
\description{
A dataset containing questions, categories, answers, and choices of ~50000 trivia questions.
}
\keyword{datasets}
