% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_as_csv.R
\name{write_as_csv}
\alias{write_as_csv}
\title{write_as_csv}
\usage{
write_as_csv(filename)
}
\arguments{
\item{filename}{Name of file to be written}
}
\value{
Writes a dataframe as a csv to file system
}
\description{
Write the list of questions as a csv to your working directory
}
\examples{
triviar::write_as_csv("trivia_questions")
}
